feat:2026.2.1.13:09:视频使用的代码版本

+++ project/code/param_config.h
@@ -10,8 +10,8 @@ extern int16_t RightPWM;						//右轮PWM
 extern int16_t AvePWM;							//平均PWM
 extern int16_t DifPWM;							//差分PWM
 
-extern int16_t Pre_LeftSpeed, Pre_RightSpeed;	//左轮上次速度，右轮上次速度
-extern int16_t LeftSpeed, RightSpeed;			//左轮实际速度，右轮实际速度
+extern float Pre_LeftSpeed, Pre_RightSpeed;	//左轮上次速度，右轮上次速度
+extern float LeftSpeed, RightSpeed;			//左轮实际速度，右轮实际速度
 extern float AveSpeed, DifSpeed;				//平均实际速度，差分实际速度
